requirements.txt
py-solc==3.2.0  
pika==1.0.0 
web3==4.8.3  